do not try to fork to the same app

+++ lib/heroku/command/fork.rb
@@ -20,6 +20,9 @@ module Heroku::Command
 
       from = app
       to = shift_argument || "#{from}-#{(rand*1000).to_i}"
+      if from == to
+        raise Heroku::Command::CommandFailed.new("Cannot fork to the same app.")
+      end
 
       from_info = api.get_app(from).body
 
@@ -91,10 +94,12 @@ module Heroku::Command
       puts "Fork complete, view it at #{to_info['web_url']}"
     rescue Exception => e
       puts "Fork failed!"
-      action("Clearing up resources") do
-        begin
-          api.delete_app(to)
-        rescue Heroku::API::Errors::NotFound
+      if to != from
+        action("Clearing up resources") do
+          begin
+            api.delete_app(to)
+          rescue Heroku::API::Errors::NotFound
+          end
         end
       end
       puts "Original exception below:"
